created new subclass enemy and had all existing enemies extend that class
